[MTM-16] [#88] chore: ApplicationClient파일 생성 후 CSR 분리
